Added validation to doesn't run kw-b command twice for production environment.

+++ gulp/tasks/pack-distro.js
@@ -4,7 +4,7 @@ var fs = require('fs');
 
 var cwd = process.cwd();
 var kwEnvConfigFile = cwd + '/cnf/environment';
-var repo = 'https://github.com/PECE-project/pece-distro.git';
+var repo = 'https://github.com/revagomes/pece-distro.git';
 var commitMessage = 'Added new release based on master of development repository.'
 
 gulp.task('pack-distro', function () {
@@ -19,14 +19,20 @@ gulp.task('pack-distro', function () {
 });
 
 var packDistroCallback = function () {
-  // Backup kw previous config and set kraftwagen env config to production
-  // in order to prevent symlinks in the profile filder.
-  shell.exec('mv ' + kwEnvConfigFile + ' ' + kwEnvConfigFile + '.bkp');
-  shell.exec('drush kw-setup-env production');
-  shell.exec('drush kw-b');
-  // Remove the updated config file and restore previous kraftwagen env config.
-  shell.exec('rm ' + kwEnvConfigFile);
-  shell.exec('mv ' + kwEnvConfigFile + '.bkp ' + kwEnvConfigFile);
+  if (!process.env.IS_PRODUCTION) {
+    // Backup kw previous config and set kraftwagen env config to production
+    // in order to prevent symlinks in the profile filder.
+    shell.exec('mv ' + kwEnvConfigFile + ' ' + kwEnvConfigFile + '.bkp');
+
+    shell.exec('drush kw-setup-env production');
+    shell.exec('drush kw-b');
+
+    // Remove the updated config file and restore previous kraftwagen env config.
+    shell.exec('rm ' + kwEnvConfigFile);
+    shell.exec('mv ' + kwEnvConfigFile + '.bkp ' + kwEnvConfigFile);
+
+    return;
+  }
 
   shell.exec('git clone ' + repo + ' distro');
